Rename files for consistant filenames
